break react components into individual files
