+++ app/addons/components/components/tabelement.js
@@ -1,0 +1,55 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy of
+// the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations under
+// the License.
+
+import React from "react";
+import ReactDOM from "react-dom";
+
+export const TabElement = ({selected, text, onChange, iconClass}) => {
+
+  const additionalClass = selected ? 'tab-element-checked' : '';
+
+  return (
+    <li className={`component-tab-element ${additionalClass}`}>
+
+      <label>
+        <div className="tab-element-indicator-wrapper">
+          <div className="tab-element-indicator"></div>
+        </div>
+        <div className="tab-element-content">
+          <i className={iconClass}></i>
+          <input
+            type="radio"
+            value={text}
+            checked={selected}
+            onChange={onChange} />
+
+          {text}
+        </div>
+      </label>
+    </li>
+
+  );
+};
+TabElement.propTypes = {
+  selected: React.PropTypes.bool.isRequired,
+  text: React.PropTypes.string.isRequired,
+  onChange: React.PropTypes.func.isRequired,
+  iconClass: React.PropTypes.string,
+};
+
+export const TabElementWrapper = ({children}) => {
+  return (
+    <ul className="component-tab-element-wrapper">
+      {children}
+    </ul>
+  );
+};
